Revert "Remove test package UI dependency cycle (#11530)"
This reverts commit 89ba28afcbd26e94c2a63ae608dff1a70210bf07.
