+++ packages/test/tsconfig.json
@@ -11,7 +11,9 @@
     "verbatimModuleSyntax": true,
     "erasableSyntaxOnly": true,
     "isolatedModules": true,
-    "skipLibCheck": true
+    "skipLibCheck": true,
+    "jsx": "react-jsx",
+    "jsxImportSource": "@remix-run/ui"
   },
   "exclude": ["dist"]
 }
